add game tests
