+++ __mocks__/client.js
@@ -1,0 +1,7 @@
+import { JSDOM } from 'jsdom';
+
+const dom = new JSDOM();
+global.document = dom.window.document;
+global.window = dom.window;
+
+exports.document = document;
